+++ src/Workspaces/Core/Desktop/Workspace/Storage/PersistenceStorageServiceFactory.cs
@@ -28,6 +28,11 @@ namespace Microsoft.CodeAnalysis.Storage
             switch (database)
             {
                 case StorageDatabase.SQLite:
+                    if (!SQLitePersistentStorageService.TryInitializeLibraries())
+                    {
+                        break;
+                    }
+
                     var locationService = workspaceServices.GetService<IPersistentStorageLocationService>();
 
                     if (locationService != null)
